+++ doc/versionHistory.rst
@@ -5,6 +5,13 @@
 ##################
 Version History
 ##################
+.. _lsst.ts.wep-14.16.0:
+
+-------------
+ 14.16.0
+-------------
+* Add DonutSizeCorrelator to measure donut diameters at the exposure level.
+* Updated handling of radius information to donut stamps metadata, and donut quality table.
 
 .. _lsst.ts.wep-14.15.0: